+++ price-feeder/router/v1/router_test.go
@@ -14,6 +14,8 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/telemetry"
 	"github.com/umee-network/umee/price-feeder/config"
+	"github.com/umee-network/umee/price-feeder/oracle"
+	"github.com/umee-network/umee/price-feeder/oracle/provider"
 	v1 "github.com/umee-network/umee/price-feeder/router/v1"
 )
 
@@ -23,6 +25,17 @@ var (
 	mockPrices = map[string]sdk.Dec{
 		"ATOM": sdk.MustNewDecFromStr("34.84"),
 		"UMEE": sdk.MustNewDecFromStr("4.21"),
+	}
+
+	mockComputedPrices = map[provider.Name]map[string]sdk.Dec{
+		provider.ProviderBinance: {
+			"ATOM": sdk.MustNewDecFromStr("28.21000000"),
+			"UMEE": sdk.MustNewDecFromStr("1.13000000"),
+		},
+		provider.ProviderKraken: {
+			"ATOM": sdk.MustNewDecFromStr("28.268700"),
+			"UMEE": sdk.MustNewDecFromStr("1.13000000"),
+		},
 	}
 )
 
@@ -34,6 +47,14 @@ func (m mockOracle) GetLastPriceSyncTimestamp() time.Time {
 
 func (m mockOracle) GetPrices() map[string]sdk.Dec {
 	return mockPrices
+}
+
+func (m mockOracle) GetTvwapPrices() oracle.PricesByProvider {
+	return mockComputedPrices
+}
+
+func (m mockOracle) GetVwapPrices() oracle.PricesByProvider {
+	return mockComputedPrices
 }
 
 type mockMetrics struct{}
@@ -101,4 +122,32 @@ func (rts *RouterTestSuite) TestPrices() {
 	rts.Require().Equal(respBody.Prices["ATOM"], mockPrices["ATOM"])
 	rts.Require().Equal(respBody.Prices["UMEE"], mockPrices["UMEE"])
 	rts.Require().Equal(respBody.Prices["FOO"], sdk.Dec{})
+}
+
+func (rts *RouterTestSuite) TestTvwap() {
+	req, err := http.NewRequest("GET", "/api/v1/prices/providers/tvwap", nil)
+	rts.Require().NoError(err)
+	response := rts.executeRequest(req)
+	rts.Require().Equal(http.StatusOK, response.Code)
+
+	var respBody v1.PricesPerProviderResponse
+	rts.Require().NoError(json.Unmarshal(response.Body.Bytes(), &respBody))
+	rts.Require().Equal(
+		respBody.Prices[provider.ProviderBinance]["ATOM"],
+		mockComputedPrices[provider.ProviderBinance]["ATOM"],
+	)
+}
+
+func (rts *RouterTestSuite) TestVwap() {
+	req, err := http.NewRequest("GET", "/api/v1/prices/providers/vwap", nil)
+	rts.Require().NoError(err)
+	response := rts.executeRequest(req)
+	rts.Require().Equal(http.StatusOK, response.Code)
+
+	var respBody v1.PricesPerProviderResponse
+	rts.Require().NoError(json.Unmarshal(response.Body.Bytes(), &respBody))
+	rts.Require().Equal(
+		respBody.Prices[provider.ProviderBinance]["ATOM"],
+		mockComputedPrices[provider.ProviderBinance]["ATOM"],
+	)
 }