+++ price-feeder/CHANGELOG.md
@@ -56,6 +56,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - [1328](https://github.com/umee-network/umee/pull/1328) Add bitget provider.
 - [1339](https://github.com/umee-network/umee/pull/1339) Add mexc provider.
+- [1445](https://github.com/umee-network/umee/pull/1445) Add computed prices api endpoints for debugging.
 - [1448](https://github.com/umee-network/umee/pull/1448) Add crypto.com provider.
 
 ### Bugs